wired in test harness, pass tests for embeds_one
create models in spec file for now, i'll extract them later.
tests require mongodb to be running on localhost default port.
the db setup and teardown is done with convenience methods in
the spec helper and will probably stay there. the embeds_one
spec reveales a bug in the Concern with the @embedded_Changes
conditional assignment. sometimes the hash is empty so the
expression ins't evaluated leaving us with an empty changes hash